requirements.txt
appnope==0.1.2
asgiref==3.4.1
autopep8==1.6.0
backcall==0.2.0
black==21.12b0
click==8.0.3
decorator==5.1.0
Deprecated==1.2.13
Django==4.0
django-extensions==3.1.5
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
environs==9.3.5
flake8==4.0.1
gunicorn==20.1.0
ipdb==0.13.9
ipython==7.30.1
jedi==0.18.1
marshmallow==3.14.1
matplotlib-inline==0.1.3
mccabe==0.6.1
mypy-extensions==0.4.3
parso==0.8.3
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.4.0
prompt-toolkit==3.0.24
psycopg2-binary==2.9.2
ptyprocess==0.7.0
pycodestyle==2.8.0
pyflakes==2.4.0
Pygments==2.10.0
PyJWT==2.3.0
python-dotenv==0.19.2
pytz==2021.3
redis==4.0.2
sqlparse==0.4.2
toml==0.10.2
tomli==1.2.3
traitlets==5.1.1
typing_extensions==4.0.1
wcwidth==0.2.5
wrapt==1.13.3
